LayoutTests/fast/encoding/char-encoding-expected.txt

This tests encoding characters in various character sets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS encode('UTF-8', 'U+00A0') is '%C2%A0'
PASS encode('macintosh', 'U+221A') is '%C3'
PASS encode('MacRoman', 'U+221A') is '%C3'
PASS successfullyParsed is true

TEST COMPLETE
